Update request view.
